Add uptime-tracker.sh at version 2.7
This script was originally written by me, but was only distributed in the
Karoshi Server project. Now it has its own repository and is licensed under
the MIT license instead of the GPLv2, so proprietary systems can use it too.

Version 2.7 is a minor update from version 2.6, with the aforementioned
license changes and the ability to create the spool directory if it does not
exist.