Track power failures
With the 'auto' mode, power failures can be tracked. Each interval the 'auto'
mode adds a '*' to the end of the uptime string, and removes it if the command
is terminated normally (normal shutdown). However, in the case of a power
failure, the character will not be removed, and can be detected by the output
commands to mean a power failure.

Future ideas - hook for supporting UPS power 'failures', where the system is
shut down safely using a UPS, but should be treated as a power failure.